requirements.txt
boto3
pyyaml

numpy
scipy
scikit-learn
pandas
Cython
pyarrow